requirements.txt
django==1.6.5
south
django-annoying
PIL